LPS-77944 Move to BundleActivator to ensure that it is initialized on time
