+++ modules/apps/wsrp/wsrp-service/bnd.bnd
@@ -1,3 +1,4 @@
+Bundle-Activator: com.liferay.wsrp.internal.activator.WSRPServiceBundleActivator
 Bundle-Name: Liferay WSRP Service
 Bundle-SymbolicName: com.liferay.wsrp.service
 Bundle-Version: 2.0.0
